+++ warehouse/src/main/java/warehouse/domain/usedgoods/converter/UsedGoodsConverter.java
@@ -3,9 +3,6 @@ package warehouse.domain.usedgoods.converter;
 import db.domain.usedgoods.EntitySearchCondition;
 import db.domain.usedgoods.UsedGoodsEntity;
 import global.annotation.Converter;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import org.springframework.data.domain.Pageable;
 import warehouse.domain.goods.controller.model.GoodsResponse;
 import warehouse.domain.usedgoods.controller.model.request.SearchCondition;
