+++ warehouse/src/main/java/warehouse/domain/usedgoods/service/UsedGoodsService.java
@@ -5,9 +5,7 @@ import db.domain.usedgoods.QueryUsedGoodsRepository;
 import db.domain.usedgoods.UsedGoodsEntity;
 import db.domain.usedgoods.UsedGoodsRepository;
 import db.domain.usedgoods.enums.UsedGoodsStatus;
-import db.domain.usedgoodsorder.UsedGoodsOrderEntity;
 import db.domain.usedgoodsorder.UsedGoodsOrderRepository;
-import db.domain.usedgoodsorder.enums.UsedGoodsOrderStatus;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,7 +14,6 @@ import org.springframework.stereotype.Service;
 import warehouse.common.error.UsedGoodsErrorCode;
 import warehouse.common.exception.usedGoods.GoodsNotInUsedStatus;
 import warehouse.common.exception.usedGoods.UsedGoodsNotFoundException;
-import warehouse.common.exception.usedGoods.UsedGoodsOrderNotFoundException;
 
 @Service
 @RequiredArgsConstructor
@@ -24,7 +21,6 @@ public class UsedGoodsService {
 
     private final UsedGoodsRepository usedGoodsRepository;
     private final QueryUsedGoodsRepository queryUsedGoodsRepository;
-    private final UsedGoodsOrderRepository usedGoodsOrderRepository;
 
     public void register(UsedGoodsEntity usedGoodsEntity) {
         usedGoodsEntity.setPostedAt(LocalDateTime.now());
@@ -67,38 +63,6 @@ public class UsedGoodsService {
             throw new UsedGoodsNotFoundException(UsedGoodsErrorCode.USED_GOODS_NOT_FOUND);
         }
         return searchList;
-    }
-
-    public UsedGoodsOrderEntity requestOrder(UsedGoodsOrderEntity orderEntity) {
-        orderEntity.setStatus(UsedGoodsOrderStatus.REGISTERED);
-        orderEntity.setCreatedAt(LocalDateTime.now());
-        return usedGoodsOrderRepository.save(orderEntity);
-    }
-
-    public List<UsedGoodsOrderEntity> getUsedGoodsOrderListBy(Long usedGoodsId) {
-        List<UsedGoodsOrderEntity> orderEntityList = usedGoodsOrderRepository.findAllByUsedGoodsId(
-            usedGoodsId);
-        if (orderEntityList.isEmpty()) {
-            throw new UsedGoodsOrderNotFoundException(
-                UsedGoodsErrorCode.USED_GOODS_ORDER_NOT_FOUND);
-        }
-        return orderEntityList;
-    }
-
-    public UsedGoodsOrderEntity getUsedGoodsOrderBy(Long usedGoodsOrderId) {
-        return usedGoodsOrderRepository.findFirstById(usedGoodsOrderId)
-            .orElseThrow(() -> new UsedGoodsOrderNotFoundException(
-                UsedGoodsErrorCode.USED_GOODS_ORDER_NOT_FOUND));
-    }
-
-    public void setUsedGoodsOrderStatusBy(UsedGoodsOrderEntity orderEntity,
-        UsedGoodsOrderStatus status) {
-        orderEntity.setStatus(status);
-        usedGoodsOrderRepository.save(orderEntity);
-    }
-
-    public Boolean hasExistingOrder(Long usedGoodsId, Long userId) {
-        return usedGoodsOrderRepository.findByUsedGoodsIdAndUserId(usedGoodsId, userId).isPresent();
     }
 
 }
